Adds to #595 move capture workload to tasks
