+++ mypy.ini
@@ -9,6 +9,10 @@ check_untyped_defs = True
 warn_unreachable = True
 strict_equality = True
 
+[mypy-posthog.tasks.*]
+disallow_untyped_defs = True
+check_untyped_defs = True
+
 [mypy.plugins.django-stubs]
 django_settings_module = posthog.settings
